[fix] Various concurrency issues
Adds session_id to FLUSH_REQ and GET_LOG_BUFFER messages to
ensure FIFO processing at the replica. Also initializes
replica_log_map and replica_index_map to NULL which is mandatory
for the utash to work. Finally fixes(maybe) an issue with the
sleep/awake mechanism of spinning thread to its workers. Spinner
checks also the size of the queue of a worker prior to sent him
for sleep.