[fix][refactor] Fixes wrong mapping from region to kreon db
Keeps Dataserver regions is an array of krm_region_desc *
instead of krm_region_desc. The previous version kept them in
an array. As a result a reference to krm_region_desc * could
probably be altered after the insertion of a new region. Also
due to bad alignment in the position of pending replica operations
in db_descriptor we had a 4x drop in server throughput. Finally,
we add intelligence to the spinning thread to wake up a worker
that is sleeping in case its queue size is > 0. (This caused
deadlocks).